+++ РГР/password.h
@@ -1,5 +1,5 @@
 #ifndef PASSWORD_H
 #define PASSWORD_H
 #include <string>
-std::string get_password(std::string key, long long p, long long q);
+std::string get_password(std::string key, int64_t  p, int64_t  q);
 #endif
