+++ РГР/rsa.cpp
@@ -11,7 +11,7 @@ bool checkIfNumericRSA(string str)
     else
         return false;
 }
-long long calculate_gcd(long long p, long long q)
+int64_t  calculate_gcd(int64_t  p, int64_t  q)
 {
     if (q == 0)
         return p;
@@ -19,9 +19,9 @@ long long calculate_gcd(long long p, long long q)
         return calculate_gcd(q, p % q);
 }
 
-long long calculate_e(long long phi)
+int64_t  calculate_e(int64_t  phi)
 {
-    long long e = 2;
+    int64_t  e = 2;
     while (e < phi && calculate_gcd(e, phi) != 1)
     {
         e++;
@@ -29,9 +29,9 @@ long long calculate_e(long long phi)
     return e;
 }
 
-long long calculate_d(long long e, long long phi)
+int64_t  calculate_d(int64_t  e, int64_t  phi)
 {
-    long long d = 1;
+    int64_t  d = 1;
     while ((d * e) % phi != 1)
     {
         d++;
@@ -39,12 +39,12 @@ long long calculate_d(long long e, long long phi)
     return d;
 }
 
-bool is_prime(long long num)
+bool is_prime(int64_t  num)
 {
     if (num == 1)
         return false;
 
-    for (long long i = 2; i <= sqrt(num); i++)
+    for (int64_t  i = 2; i <= sqrt(num); i++)
     {
         if (num % i == 0)
             return false;
@@ -53,7 +53,7 @@ bool is_prime(long long num)
     return true;
 }
 
-bool can_encrypt1(long long p, long long q)
+bool can_encrypt1(int64_t  p, int64_t  q)
 {
     if ((p-1) * (q-1) >= 255)
         return true;
@@ -61,7 +61,7 @@ bool can_encrypt1(long long p, long long q)
         return false;
 }
 
-bool is_different(long long p, long long q)
+bool is_different(int64_t  p, int64_t  q)
 {
     if (p != q)
         return true;
@@ -69,7 +69,7 @@ bool is_different(long long p, long long q)
         return false;
 }
 
-bool can_encrypt2(long long p, long long q)
+bool can_encrypt2(int64_t  p, int64_t  q)
 {
     if (p*q <= 3025550009)
         return true;
@@ -77,7 +77,7 @@ bool can_encrypt2(long long p, long long q)
         return false;
 }
 
-void get_primes(long long& p, long long& q)
+void get_primes(int64_t & p, int64_t & q)
 {
     string primes;
     cout << "Enter two different prime numbers separated by space: ";
@@ -120,9 +120,9 @@ void get_primes(long long& p, long long& q)
     }
 }
 
-long long fast_pow(long long a, long long b, long long n)
+int64_t  fast_pow(int64_t  a, int64_t  b, int64_t  n)
 {
-    long long res = 1;
+    int64_t  res = 1;
     while (b > 0)
     {
         if (b & 1)
@@ -140,20 +140,20 @@ void rsa_encrypt(std::string inputFileName, std::string outputFileName)
     wifstream inputFile(inputFileName);
     wofstream outputFile(outputFileName);
     wstring line;
-    long long p, q;
+    int64_t  p, q;
     get_primes(p, q);
-    long long n = p * q;
-    long long phi = (p - 1) * (q - 1);
-    long long e = calculate_e(phi);
-    long long d = calculate_d(e, phi);
+    int64_t  n = p * q;
+    int64_t  phi = (p - 1) * (q - 1);
+    int64_t  e = calculate_e(phi);
+    int64_t  d = calculate_d(e, phi);
     cout << "Public key: " << e << " ; " << n << endl;
     cout << "Private key: " << d << " ; " << n << endl;
     while (getline(inputFile, line))
     {
         for (wchar_t& symbol : line)
         {
-            long long m = symbol;
-            long long crypted_m = fast_pow(m, e, n);
+            int64_t  m = symbol;
+            int64_t  crypted_m = fast_pow(m, e, n);
             outputFile << crypted_m << " ";
         }
         outputFile << "\n";
@@ -168,22 +168,22 @@ void rsa_decrypt(std::string inputFileName, std::string outputFileName)
     ifstream inputFile(inputFileName);
     ofstream outputFile(outputFileName);
     string line;
-    long long p, q;
+    int64_t  p, q;
     get_primes(p, q);
-    long long n = p * q;
-    long long phi = (p - 1) * (q - 1);
-    long long e = calculate_e(phi);
-    long long d = calculate_d(e, phi);
+    int64_t  n = p * q;
+    int64_t  phi = (p - 1) * (q - 1);
+    int64_t  e = calculate_e(phi);
+    int64_t  d = calculate_d(e, phi);
     while (getline(inputFile, line))
     {
         string crypted_m_str;
-        long long crypted_m;
+        int64_t  crypted_m;
         stringstream ss(line);
         while (getline(ss, crypted_m_str, ' '))
         {
             if (checkIfNumericRSA(crypted_m_str)) {
                 crypted_m = stoull(crypted_m_str);
-                long long m = fast_pow(crypted_m, d, n);
+                int64_t  m = fast_pow(crypted_m, d, n);
                 char symbol = (char)m;
                 outputFile << symbol;
             }
